pyproject.toml
[project]
name = "photo-intel"
version = "0.1.0"
description = "Self-hosted photo intelligence and digest system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pillow>=10.3.0",
    "python-magic>=0.4.27",
    "httpx>=0.27.0",
    "watchfiles>=0.22.0",
    "pytesseract>=0.3.10",
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "exifread>=3.0.0",
    "apscheduler>=3.10.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
photo-intel = "src.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
